Cpu continue

+++ includes/modules/Cpu.hpp
@@ -5,7 +5,7 @@
 // Login   <lacomm_m@epitech.net>
 // 
 // Started on  Sat Jan 21 16:57:15 2017 Manon Lacommare
-// Last update Sun Jan 22 06:22:10 2017 Manon Lacommare
+// Last update Sun Jan 22 07:07:20 2017 Manon Lacommare
 //
 
 #ifndef CPU_HPP_
@@ -33,7 +33,7 @@ class			Cpu : public IMonitorModule
   double		val;
 
 public:
-  std::vector<double>	cpu;
+  std::vector<int>	cpu;
 
   Cpu();
   virtual ~Cpu();
